Refactor serializers and restructure module organization
Modularized serializers into distinct files for improved clarity and maintainability. Refactored placeholder caching, language, and page handling utilities into dedicated modules. Adjusted settings and base views to align with the new structure.
