+++ djangocms_rest/serializers/utils/cache.py
