+++ djangocms_rest/urls.py
@@ -1,15 +1,15 @@
 from django.urls import path
 
-from . import views
+from djangocms_rest.views import LanguageListView, PageDetailView, PageTreeListView, PlaceholderDetailView
 
 urlpatterns = [
-    path("languages/", views.LanguageListView.as_view(), name="cms-language-list"),
-    path("<slug:language>/pages-tree/", views.PageTreeListView.as_view(), name="cms-page-tree-list"),
-    path("<slug:language>/pages-root/", views.PageDetailView.as_view(), name="cms-page-root"),
-    path("<slug:language>/pages/<path:path>/", views.PageDetailView.as_view(), name="cms-page-detail"),
+    path("languages/", LanguageListView.as_view(), name="cms-language-list"),
+    path("<slug:language>/pages-tree/", PageTreeListView.as_view(), name="cms-page-tree-list"),
+    path("<slug:language>/pages-root/", PageDetailView.as_view(), name="cms-page-root"),
+    path("<slug:language>/pages/<path:path>/", PageDetailView.as_view(), name="cms-page-detail"),
     path(
         "<slug:language>/placeholders/<int:content_type_id>/<int:object_id>/<str:slot>/",
-        views.PlaceholderDetailView.as_view(),
+        PlaceholderDetailView.as_view(),
         name="cms-placeholder-detail",
     ),
 ]
